added Edit button, refactored function a little

+++ scripts/main.js
@@ -88,7 +88,7 @@ let createCard = function(itemObj, itemElem) {
 
   cardTitle.textContent = itemObj.title;
   cardContentDiv.appendChild(itemElem);
-  cardAction1.textContent = `Delay`;
+  cardAction1.textContent = `Edit`;
   cardAction2.textContent = "Tweet";
   cardAction3.textContent = "Delete";
   cardAction2.addEventListener('click', tweet);
@@ -100,6 +100,10 @@ let createCard = function(itemObj, itemElem) {
     cta3: cardAction3
   };
 
+  cardObj.cta1.addEventListener('click', function(e) {
+    editData(itemObj, cardObj);
+  })
+
   cardObj.cta3.addEventListener('click', function(event) {
     deleteData(itemObj.id);
     cardObj.card.remove();
@@ -107,6 +111,74 @@ let createCard = function(itemObj, itemElem) {
 
   return cardObj;
 };
+
+var editData = function(itemObj, cardObj) {
+  if (itemObj.type === 'text') {
+    editText(itemObj, cardObj);
+  } else 
+  if (itemObj.type === 'image') {
+    editImage(itemObj, cardObj);
+  } else
+  if (itemObj.type === 'list') {
+    editList(itemObj, cardObj);
+  }
+  else ( console.log('smth wrong'))
+}
+
+var modalTrigger = document.querySelector('ul.right>li>a.modal-trigger');
+
+var editText = function(itemObj, cardObj) {
+  modalTrigger.click();
+  addTextModal.click();
+  var title = document.querySelector('input.title-input');
+  title.value = itemObj.title;
+  var text = document.querySelector('textarea');
+  text.textContent = itemObj.content
+  var buttonAdd = document.querySelector('#add');
+  buttonAdd.addEventListener('click', function() {
+    deleteData(itemObj.id);
+    cardObj.card.remove();
+    //add position
+  })
+}
+
+var editImage = function(itemObj, cardObj) {
+  modalTrigger.click();
+  addImageModal.click();
+  var title = document.querySelector('input.title-input');
+  title.value = itemObj.title;
+  var urlForm = document.querySelector('#url-form');
+  urlForm.value = itemObj.content;
+  var buttonAdd = document.querySelector('#add');
+  buttonAdd.addEventListener('click', function() {
+    deleteData(itemObj.id);
+    cardObj.card.remove();
+    //add position
+  })
+}
+
+var editList = function(itemObj, cardObj) {
+  modalTrigger.click();
+  addListModal.click();
+  var title = document.querySelector('input.title-input');
+  title.value = itemObj.title;
+  var bulletPoint = document.querySelector('input.bullet-point');
+  bulletPoint.focus();
+  var ul = document.querySelector('div.list-form>ul.bulletpoints')
+  for (var i = 0; i < itemObj.content.length; i++) {
+    if (itemObj.content[i] !== '') {
+      document.activeElement.value = itemObj.content[i];
+      appendInputLi(ul);
+    }
+  }
+
+  var buttonAdd = document.querySelector('#add');
+  buttonAdd.addEventListener('click', function() {
+    deleteData(itemObj.id);
+    cardObj.card.remove();
+    //add position
+  })
+}
 
 ///#####################
 
@@ -159,7 +231,7 @@ var addTextModal = document.querySelector('#add-text');
 addTextModal.addEventListener('click', function() {
 
   var divForm = createTextForm();
-  var divButtons = createButtons('text', divForm, null)
+  var divButtons = createButtons('text')
 
   divForm.appendChild(divButtons);
   divModalContent.appendChild(divForm);
@@ -198,7 +270,7 @@ var addImageModal = document.querySelector('#add-image');
 addImageModal.addEventListener('click', function() {
 
   var divForm = createImageForm();
-  var divButtons = createButtons('image', divForm, null);
+  var divButtons = createButtons('image');
   divForm.appendChild(divButtons);
   divModalContent.appendChild(divForm);
 })
@@ -240,7 +312,8 @@ var deleteElement = function(element) {
 var createAddBtn = function() {
   var submitButton = document.createElement('button');
   submitButton.textContent = 'Add';
-  submitButton.className = 'cancel-submit-btn'
+  submitButton.className = 'cancel-submit-btn';
+  submitButton.id = 'add';
   return submitButton;
 }
 
@@ -248,6 +321,7 @@ var createCancelBtn = function() {
   var cancelButton = document.createElement('button');
   cancelButton.textContent = 'Cancel';
   cancelButton.className = 'cancel-submit-btn';
+  cancelButton.id = 'cancel';
   return cancelButton;
 }
 
@@ -262,22 +336,21 @@ var placeFirst = function(element) {
 var addListModal = document.querySelector('#add-list');
 
 addListModal.addEventListener('click', function() {
-  var arrDOMElements = []
-
   var divWrapper = document.createElement('div');
+  divWrapper.className = 'div-form';
   var divForm = document.createElement('div');
   divForm.className = "list-form";
 
   var ul = document.createElement('ul')
   ul.className = 'bulletpoints';
-  appendInputLi(ul, arrDOMElements);
+  appendInputLi(ul);
 
   var titleInput = createTitleInput();
 
   divForm.appendChild(titleInput);
   divForm.appendChild(ul);
 
-  var divButtons = createButtons('list', divWrapper, arrDOMElements);
+  var divButtons = createButtons('list');
   divWrapper.appendChild(divForm);
   divWrapper.appendChild(divButtons);
   divModalContent.appendChild(divWrapper);
@@ -285,22 +358,24 @@ addListModal.addEventListener('click', function() {
 
   divForm.addEventListener('keydown', function(e) {
     if(e.keyCode === 13 && e.target.className === 'title-input') {
+      var bullet = document.querySelector('input.bullet-point');
       bullet.focus();
+      
     }
     if(e.keyCode === 13 && e.target.className === 'bullet-point') {
       e.preventDefault();
-      appendInputLi(ul, arrDOMElements);
+      appendInputLi(ul);
     }
   })
+  
 })
 
-var appendInputLi = function(ul, arrOfElements) {
+var appendInputLi = function(ul) {
   var li = document.createElement('li');
   var newBullet = createBulletInput();
   li.appendChild(newBullet)
   ul.appendChild(li);
   newBullet.focus();
-  arrOfElements.push(newBullet);
 }
 
 var createBulletInput = function() {
@@ -327,7 +402,8 @@ var createListElements = function(listObj) {
 }
 
 
-var createListObj = function(arrDOMElements) {
+var createListObj = function() {
+  var arrDOMElements = document.querySelectorAll('input.bullet-point')
   var titleInput = document.querySelector('div.list-form>input');
   var listContents = [];
   for (var i = 0; i < arrDOMElements.length; i ++) {
@@ -341,8 +417,9 @@ var createListObj = function(arrDOMElements) {
   return listObj;
 }
 
-var createButtons = function(objType, divToDelete, arrDOMElements ) {
 
+
+var createButtons = function(objType ) {
   var divButtons = document.createElement('div');
   divButtons.className = 'buttons';
   var buttonAdd = createAddBtn();
@@ -356,7 +433,7 @@ var createButtons = function(objType, divToDelete, arrDOMElements ) {
       someObj = createTextObj();
     } else
     if (objType === 'list') {
-      someObj = createListObj(arrDOMElements);
+      someObj = createListObj();
     } else 
     if (objType === 'image') {
       someObj = createImgObj();
@@ -364,7 +441,7 @@ var createButtons = function(objType, divToDelete, arrDOMElements ) {
       console.log('check object type!! createButtons function')
       return
     }
-
+    var divToDelete = document.querySelector('div.div-form');
     writeData(someObj).then(testfunc);
     var cardObj = load(someObj);
     placeFirst(cardObj.card);
@@ -373,8 +450,34 @@ var createButtons = function(objType, divToDelete, arrDOMElements ) {
   })
 
   buttonCancel.addEventListener('click', function(){
-    deleteElement(divToDelete);
+    var divToDelete = document.querySelector('div.div-form');
+    if (divToDelete) {
+      deleteElement(divToDelete);
+    }
+    
   })
+
+  var modalOverlayArea = document.querySelector('div.modal-overlay');
+  document.addEventListener('click', function(e) {
+    if (e.target === modalOverlayArea) {
+      var divToDelete = document.querySelector('div.div-form');
+      if (divToDelete) {
+        deleteElement(divToDelete);
+      }
+      
+      
+    }
+  })
+
+  var closeButton = document.querySelector('div.modal-footer>a.modal-close')
+  closeButton.addEventListener('click', function() {
+    var divToDelete = document.querySelector('div.div-form');
+    if (divToDelete) {
+      deleteElement(divToDelete);
+    }
+    
+  })
+
   return divButtons;
 }
 
@@ -384,7 +487,7 @@ var speechModal = document.querySelector('#speech');
 
 speechModal.addEventListener('click', function() {
   var divForm = createSpeechForm();
-  var divButtons = createButtons('text', divForm, null)
+  var divButtons = createButtons('text')
 
   divForm.appendChild(divButtons);
   divModalContent.appendChild(divForm);
@@ -400,7 +503,8 @@ speechModal.addEventListener('click', function() {
     },
     'cancel': function() {
       annyang.abort();
-      deleteElement(divForm);
+      var divToDelete = document.querySelector('div.div-form');
+      deleteElement(divToDelete);
     },
     'start over': function() {
       speechTextArea.textContent = '';
@@ -417,13 +521,13 @@ speechModal.addEventListener('click', function() {
 })
 
 
-var initRecognition = function() {
-var recognition = new (window.SpeechRecognition || window.webkitSpeechRecognition || window.mozSpeechRecognition || window.msSpeechRecognition)();
-recognition.lang = 'en-US';
-recognition.interimResults = false;
-recognition.maxAlternatives = 5;
-return recognition;
-}
+// var initRecognition = function() {
+// var recognition = new (window.SpeechRecognition || window.webkitSpeechRecognition || window.mozSpeechRecognition || window.msSpeechRecognition)();
+// recognition.lang = 'en-US';
+// recognition.interimResults = false;
+// recognition.maxAlternatives = 5;
+// return recognition;
+// }
 
 var createSpeechForm = function() {
   var divForm = document.createElement('div');
